+++ playwright/tests/titleSearch.spec.ts
@@ -9,7 +9,6 @@ When("I change the dropdown to title", async function (this: CustomWorld) {
 
 When("I input a title search term", async function (this: CustomWorld) {
   await this.page.locator("[aria-label='Item Search']").fill("IBM 1401");
-  await this.page.locator("#searchbar-button-search-bar").click();
 });
 
 Then(