+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGE LOG
 
+## [Pre-release]
+- Refactored Playwright/Cucumber tests to use new click action function
+
 ## [0.17.0]
 - Replace Search Examples with Recently Added Collections to landing page
 - Add primary hero to the landing page